Added a temporary / patchy fix for metamask onboarding bug
The bug happens when installing (or enabling) the metamask extension through the use of the Metamask Onboarding library. The page is reloaded after enabling the extension and it becomes unresponsive (sending messages to parent), where the website is unable to communicate with the extension.

Refreshing the page after a defined timeout (i.e. 5 seconds) fixes this
issue for all users.